Исправил баги из-за которых не работала регистрация и авторизация. Написал тесты.

+++ src/main/java/com/file/storage/controller/ResourceController.java
@@ -12,7 +12,6 @@ import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 
@@ -29,13 +28,14 @@ public class ResourceController {
     @GetMapping({"/resource", "/directory"})
     public ResponseEntity<?> getResourceOrDirectory(@RequestParam String path,
                                                     @AuthenticationPrincipal UserDetails userDetails) {
+        if (userDetails == null) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(new ErrorResponse("The user is not authorized")); //401
+        }
         try {
             ResourceInfoResponse resourceInfoResponse = resourceService.getResourceInfo(userDetails.getUsername(), path);
             return ResponseEntity.ok(resourceInfoResponse);
         } catch (InvalidPathException e) {
             return ResponseEntity.badRequest().body(new ErrorResponse("Invalid or missing path")); //400
-        } catch (UsernameNotFoundException e) {
-            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(new ErrorResponse("The user is not authorized")); //401
         } catch (ResourceNotFoundException e) {
             return ResponseEntity.status(HttpStatus.NOT_FOUND).body(new ErrorResponse("The resource or directory was not found")); //404
         } catch (Exception e) {
@@ -46,13 +46,14 @@ public class ResourceController {
     @DeleteMapping("/resource")
     public ResponseEntity<?> deleteResource(@RequestParam String path,
                                             @AuthenticationPrincipal UserDetails userDetails) {
+        if (userDetails == null) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(new ErrorResponse("The user is not authorized")); //401
+        }
         try {
             resourceService.deleteResource(userDetails.getUsername(), path);
             return ResponseEntity.noContent().build();
         } catch (InvalidPathException e) {
             return ResponseEntity.badRequest().body(new ErrorResponse("Invalid or missing path")); //400
-        } catch (UsernameNotFoundException e) {
-            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(new ErrorResponse("The user is not authorized")); //401
         } catch (ResourceNotFoundException e) {
             return ResponseEntity.status(HttpStatus.NOT_FOUND).body(new ErrorResponse("The resource was not found")); //404
         } catch (Exception e) {
@@ -63,6 +64,9 @@ public class ResourceController {
     @GetMapping("/resource/download")
     public ResponseEntity<?> downloadResource(@RequestParam String path,
                                               @AuthenticationPrincipal UserDetails userDetails) {
+        if (userDetails == null) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(new ErrorResponse("The user is not authorized")); //401
+        }
         try {
             InputStream inputStream = resourceService.downloadResource(userDetails.getUsername(), path);
 
@@ -76,8 +80,6 @@ public class ResourceController {
                     .body(new InputStreamResource(inputStream));
         } catch (InvalidPathException e) {
             return ResponseEntity.badRequest().body(new ErrorResponse("Invalid or missing path")); //400
-        } catch (UsernameNotFoundException e) {
-            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(new ErrorResponse("The user is not authorized")); //401
         } catch (ResourceNotFoundException e) {
             return ResponseEntity.status(HttpStatus.NOT_FOUND).body(new ErrorResponse("The resource was not found")); //404
         } catch (Exception e) {
@@ -89,13 +91,14 @@ public class ResourceController {
     public ResponseEntity<?> moveResource(@RequestParam String from,
                                           @RequestParam String to,
                                           @AuthenticationPrincipal UserDetails userDetails) {
+        if (userDetails == null) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(new ErrorResponse("The user is not authorized")); //401
+        }
         try {
             ResourceInfoResponse resourceInfoResponse = resourceService.moveResource(userDetails.getUsername(), from, to);
             return ResponseEntity.ok(resourceInfoResponse);
         } catch (InvalidPathException e) {
             return ResponseEntity.badRequest().body(new ErrorResponse("Invalid or missing path")); //400
-        } catch (UsernameNotFoundException e) {
-            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(new ErrorResponse("The user is not authorized")); //401
         } catch (ResourceNotFoundException e) {
             return ResponseEntity.status(HttpStatus.NOT_FOUND).body(new ErrorResponse("The resource was not found")); //404
         } catch (ResourceAlreadyExistsException e) {
@@ -108,13 +111,14 @@ public class ResourceController {
     @GetMapping("/resource/search")
     public ResponseEntity<?> searchResource(@RequestParam String query,
                                             @AuthenticationPrincipal UserDetails userDetails) {
+        if (userDetails == null) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(new ErrorResponse("The user is not authorized")); //401
+        }
         try {
             List<ResourceInfoResponse> resourceInfoResponse = resourceService.searchResource(userDetails.getUsername(), query);
             return ResponseEntity.ok(resourceInfoResponse);
         } catch (InvalidPathException e) {
             return ResponseEntity.badRequest().body(new ErrorResponse("Invalid or missing path")); //400
-        } catch (UsernameNotFoundException e) {
-            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(new ErrorResponse("The user is not authorized")); //401
         } catch (Exception e) {
             return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(new ErrorResponse("Internal server error")); //500
         }
@@ -124,13 +128,14 @@ public class ResourceController {
     public ResponseEntity<?> uploadResource(@RequestParam String path,
                                             @RequestParam("file") List<MultipartFile> files,
                                             @AuthenticationPrincipal UserDetails userDetails) {
+        if (userDetails == null) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(new ErrorResponse("The user is not authorized")); //401
+        }
         try {
             List<ResourceInfoResponse> resourceInfoResponse = resourceService.uploadResource(userDetails.getUsername(), path, files);
             return ResponseEntity.status(HttpStatus.CREATED).body(resourceInfoResponse);
         } catch (InvalidPathException e) {
             return ResponseEntity.badRequest().body(new ErrorResponse("Invalid or missing path")); //400
-        } catch (UsernameNotFoundException e) {
-            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(new ErrorResponse("The user is not authorized")); //401
         } catch (ResourceAlreadyExistsException e) {
             return ResponseEntity.status(HttpStatus.CONFLICT).body(new ErrorResponse("The file already exists")); //409
         } catch (Exception e) {
@@ -142,13 +147,14 @@ public class ResourceController {
     public ResponseEntity<?> createDirectory(@RequestParam String path,
                                              @AuthenticationPrincipal UserDetails userDetails) {
 
+        if (userDetails == null) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(new ErrorResponse("The user is not authorized")); //401
+        }
         try {
             ResourceInfoResponse resourceInfoResponse = resourceService.createDirectory(userDetails.getUsername(), path);
             return ResponseEntity.status(HttpStatus.CREATED).body(resourceInfoResponse);
         } catch (InvalidPathException e) {
             return ResponseEntity.badRequest().body(new ErrorResponse("Invalid or missing path")); //400
-        } catch (UsernameNotFoundException e) {
-            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(new ErrorResponse("The user is not authorized")); //401
         } catch (ResourceNotFoundException e) {
             return ResponseEntity.status(HttpStatus.NOT_FOUND).body(new ErrorResponse("The parent folder does not exist")); //404
         } catch (ResourceAlreadyExistsException e) {
